Add full contents of Japanese Curry recipe
Similar to the previous commit, I added the full components of the
recipe list of the Japanese Curry: heading, image, description,
ingredients, and steps to make it.

For ingredients, it is longer than usual for it required to
separate list of ingredients: one for the curry powder, and
another for the curry paste.