Group recipe page components to bordered sections
It would be a visual improvement if the user could easily navigate to
areas of the recipe list where the ingredients and recipes are located.
Hence, I placed them in a separate container which is styled such that
every container inside of that container has a separate border,
therefore creating a new section for users to pay attention to.